feature: refactor commodity table

+++ trade-commodity/commodity-db/src/main/java/org/study/trade/commodity/mapper/CommodityDataMapper.java
@@ -1,22 +1,22 @@
 package org.study.trade.commodity.mapper;
 
 import org.apache.ibatis.annotations.Param;
-import org.study.trade.commodity.data.CommodityData;
+import org.study.trade.commodity.mapper.data.CommodityData;
 
 import java.util.List;
 
 public interface CommodityDataMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insertSnapshot(CommodityData record);
 
     int insertSelective(CommodityData record);
 
-    CommodityData selectByPrimaryKey(Integer id);
+    CommodityData selectByPrimaryKey(Long id);
 
-    CommodityData selectByIdAndVersion(@Param("id") Integer id, @Param("version") Long version);
+    CommodityData selectByIdAndVersion(@Param("id") Long id, @Param("version") Long version);
 
-    List<CommodityData> batchByPrimaryKey(@Param("idList") List<Integer> idList);
+    List<CommodityData> batchByPrimaryKey(@Param("idList") List<Long> idList);
 
     int updateByPrimaryKeySelective(CommodityData record);
